requirements.txt
Django==1.7.2
django-db-parti==0.3.3
djangorestframework==3.0.2
djorm-ext-pgfulltext==0.10
ipython==2.3.1
oauthlib==0.7.2
pbr==0.10.7
psycopg2==2.5.4
PyJWT==0.4.0
python-memcached==1.53
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
stevedore==1.2.0
virtualenv==12.0.5
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
django-registration==1.0
